src/test/java/ua/com/foxminded/university/dao/mappers/GroupMapperTest.java
package ua.com.foxminded.university.dao.mappers;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
import ua.com.foxminded.university.dao.DaoTestConfig;
import ua.com.foxminded.university.entities.Group;

import java.sql.ResultSet;
import java.sql.SQLException;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
@SpringJUnitConfig(DaoTestConfig.class)
class GroupMapperTest {
    @Mock
    private ResultSet resultSet;

    @InjectMocks
    @Autowired
    private GroupMapper groupMapper;

    private final Group expectedGroup = new Group(1, "AC-02");

    @Test
    void shouldReturnGroupWithCorrectSettings() throws SQLException {
        when(resultSet.getInt("id")).thenReturn(1);
        when(resultSet.getString("name")).thenReturn("AC-02");

        Group actualGroup = groupMapper.mapRow(resultSet, 1);

        assertEquals(expectedGroup, actualGroup);
    }
}